+++ pkg/interactive/metaquery/handler_search_path.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/spf13/viper"
 	"github.com/turbot/go-kit/helpers"
-	pconstants "github.com/turbot/pipe-fittings/constants"
-	"github.com/turbot/pipe-fittings/querydisplay"
+	pconstants "github.com/turbot/pipe-fittings/v2/constants"
+	"github.com/turbot/pipe-fittings/v2/querydisplay"
 	"github.com/turbot/steampipe/pkg/constants"
 )