+++ pkg/interactive/metaquery/handlers.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	typeHelpers "github.com/turbot/go-kit/types"
-	pconstants "github.com/turbot/pipe-fittings/constants"
+	pconstants "github.com/turbot/pipe-fittings/v2/constants"
 	"github.com/turbot/steampipe/pkg/cmdconfig"
 	"github.com/turbot/steampipe/pkg/constants"
 	"golang.org/x/exp/maps"