+++ pkg/pluginmanager_service/plugin_manager_notifications.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"log"
 
-	"github.com/turbot/pipe-fittings/error_helpers"
+	"github.com/turbot/pipe-fittings/v2/error_helpers"
 	"github.com/turbot/steampipe/pkg/db/db_local"
 	"github.com/turbot/steampipe/pkg/steampipeconfig"
 )